+++ src/Authority.php
@@ -14,23 +14,17 @@ use Innmind\Url\Authority\{
  */
 final class Authority
 {
-    private UserInformation $userInformation;
-    private Host $host;
-    private Port $port;
-
     private function __construct(
-        UserInformation $userInformation,
-        Host $host,
-        Port $port,
+        private UserInformation $userInformation,
+        private Host $host,
+        private Port $port,
     ) {
-        $this->userInformation = $userInformation;
-        $this->host = $host;
-        $this->port = $port;
     }
 
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(
         UserInformation $userInformation,
         Host $host,
@@ -42,6 +36,7 @@ final class Authority
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function none(): self
     {
         return new self(
@@ -51,6 +46,7 @@ final class Authority
         );
     }
 
+    #[\NoDiscard]
     public function equals(self $authority): bool
     {
         return $this->userInformation->equals($authority->userInformation()) &&
@@ -58,51 +54,61 @@ final class Authority
             $this->port->equals($authority->port());
     }
 
+    #[\NoDiscard]
     public function userInformation(): UserInformation
     {
         return $this->userInformation;
     }
 
+    #[\NoDiscard]
     public function withUserInformation(UserInformation $info): self
     {
         return new self($info, $this->host, $this->port);
     }
 
+    #[\NoDiscard]
     public function withoutUserInformation(): self
     {
         return new self(UserInformation::none(), $this->host, $this->port);
     }
 
+    #[\NoDiscard]
     public function host(): Host
     {
         return $this->host;
     }
 
+    #[\NoDiscard]
     public function withHost(Host $host): self
     {
         return new self($this->userInformation, $host, $this->port);
     }
 
+    #[\NoDiscard]
     public function withoutHost(): self
     {
         return new self($this->userInformation, Host::none(), $this->port);
     }
 
+    #[\NoDiscard]
     public function port(): Port
     {
         return $this->port;
     }
 
+    #[\NoDiscard]
     public function withPort(Port $port): self
     {
         return new self($this->userInformation, $this->host, $port);
     }
 
+    #[\NoDiscard]
     public function withoutPort(): self
     {
         return new self($this->userInformation, $this->host, Port::none());
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->userInformation->format($this->host).$this->port->format();