+++ src/Authority/Port.php
@@ -3,45 +3,44 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority;
 
-use Innmind\Url\Exception\DomainException;
-
 /**
  * @psalm-immutable
  */
 final class Port
 {
-    private ?int $value;
-
-    private function __construct(?int $value)
+    private function __construct(private ?int $value)
     {
-        if (\is_int($value) && $value < 0) {
-            throw new DomainException((string) $value);
-        }
-
-        $this->value = $value;
     }
 
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(int $value): self
     {
+        if ($value < 0) {
+            throw new \DomainException((string) $value);
+        }
+
         return new self($value);
     }
 
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function none(): self
     {
         return new self(null);
     }
 
+    #[\NoDiscard]
     public function equals(self $port): bool
     {
         return $this->value === $port->value;
     }
 
+    #[\NoDiscard]
     public function format(): string
     {
         if ($this->value === null) {
@@ -51,6 +50,7 @@ final class Port
         return ':'.$this->value;
     }
 
+    #[\NoDiscard]
     public function value(): int
     {
         return match ($this->value) {
@@ -59,6 +59,7 @@ final class Port
         };
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         if ($this->value === null) {