+++ src/Authority/UserInformation/Password.php
@@ -3,10 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority\UserInformation;
 
-use Innmind\Url\Exception\{
-    DomainException,
-    PasswordCannotBeSpecifiedWithoutAUser,
-};
 use Innmind\Immutable\Str;
 
 /**
@@ -26,10 +22,11 @@ final class Password
     /**
      * @psalm-pure
      */
-    public static function of(string $value): self
+    #[\NoDiscard]
+    public static function of(#[\SensitiveParameter] string $value): self
     {
         if (!Str::of($value)->matches(self::PATTERN)) {
-            throw new DomainException($value);
+            throw new \DomainException;
         }
 
         return new self($value);
@@ -38,16 +35,19 @@ final class Password
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function none(): self
     {
         return new self('');
     }
 
+    #[\NoDiscard]
     public function equals(self $password): bool
     {
         return $this->value->getValue() === $password->value->getValue();
     }
 
+    #[\NoDiscard]
     public function format(User $user): string
     {
         if ($this->value->getValue() === '') {
@@ -55,12 +55,13 @@ final class Password
         }
 
         if ($user->toString() === '') {
-            throw new PasswordCannotBeSpecifiedWithoutAUser;
+            throw new \DomainException('Password cannot be specified without a user');
         }
 
         return $user->toString().':'.(string) $this->value->getValue();
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value->getValue();