+++ src/Authority/UserInformation/User.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority\UserInformation;
 
-use Innmind\Url\{
-    Authority\Host,
-    Exception\DomainException,
-};
+use Innmind\Url\Authority\Host;
 use Innmind\Immutable\Str;
 
 /**
@@ -15,20 +12,19 @@ use Innmind\Immutable\Str;
 final class User
 {
     private const PATTERN = '/^[\pL\pN-]+$/';
-    private string $value;
 
-    private function __construct(string $value)
+    private function __construct(private string $value)
     {
-        $this->value = $value;
     }
 
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(string $value): self
     {
         if (!Str::of($value)->matches(self::PATTERN)) {
-            throw new DomainException($value);
+            throw new \DomainException($value);
         }
 
         return new self($value);
@@ -37,16 +33,19 @@ final class User
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function none(): self
     {
         return new self('');
     }
 
+    #[\NoDiscard]
     public function equals(self $user): bool
     {
         return $this->value === $user->value;
     }
 
+    #[\NoDiscard]
     public function format(Host $host, Password $password): string
     {
         if ($this->value === '') {
@@ -56,6 +55,7 @@ final class User
         return $password->format($this).'@'.$host->toString();
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value;