+++ src/Path.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Url;
 
-use Innmind\Url\Exception\DomainException;
 use Innmind\Immutable\Str;
 
 /**
@@ -11,21 +10,20 @@ use Innmind\Immutable\Str;
  */
 abstract class Path
 {
-    private const PATTERN = '~\S+~';
-    private string $value;
+    private const string PATTERN = '~\S+~';
 
-    final private function __construct(string $value)
+    final private function __construct(private string $value)
     {
-        $this->value = $value;
     }
 
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     final public static function of(string $value): self
     {
         if (!Str::of($value)->matches(self::PATTERN)) {
-            throw new DomainException($value);
+            throw new \DomainException($value);
         }
 
         return $value[0] === '/' ? new AbsolutePath($value) : new RelativePath($value);
@@ -34,11 +32,13 @@ abstract class Path
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     final public static function none(): self
     {
         return new AbsolutePath('');
     }
 
+    #[\NoDiscard]
     final public function equals(self $path): bool
     {
         return $this->value === $path->value;
@@ -46,11 +46,13 @@ abstract class Path
 
     abstract public function absolute(): bool;
 
+    #[\NoDiscard]
     final public function directory(): bool
     {
         return $this->toString()[-1] === '/';
     }
 
+    #[\NoDiscard]
     final public function resolve(self $path): self
     {
         if ($path->absolute()) {
@@ -66,6 +68,7 @@ abstract class Path
         return self::of($parent.'/'.$path->toString());
     }
 
+    #[\NoDiscard]
     final public function format(Query $query, Fragment $fragment): string
     {
         $end = $query->format().$fragment->format();
@@ -77,6 +80,7 @@ abstract class Path
         return $this->value.$end;
     }
 
+    #[\NoDiscard]
     final public function toString(): string
     {
         return $this->value === '' ? '/' : $this->value;