+++ src/Query.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Url;
 
-use Innmind\Url\Exception\DomainException;
 use Innmind\Immutable\Str;
 use League\Uri\QueryString;
 
@@ -12,13 +11,13 @@ use League\Uri\QueryString;
  */
 final class Query
 {
-    private const PATTERN = '/^\S+$/';
+    private const string PATTERN = '/^\S+$/';
     private string $value;
 
     private function __construct(string $value)
     {
         if (!Str::of($value)->matches(self::PATTERN)) {
-            throw new DomainException($value);
+            throw new \DomainException($value);
         }
 
         $this->value = $value;
@@ -27,11 +26,12 @@ final class Query
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(string $value): self
     {
         try {
             return new self($value);
-        } catch (DomainException $e) {
+        } catch (\DomainException $e) {
             /** @psalm-suppress ImpureMethodCall */
             return new self(
                 (string) QueryString::build(QueryString::parse($value)),
@@ -42,6 +42,7 @@ final class Query
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function none(): self
     {
         $self = new self('void');
@@ -50,11 +51,13 @@ final class Query
         return $self;
     }
 
+    #[\NoDiscard]
     public function equals(self $query): bool
     {
         return $this->value === $query->value;
     }
 
+    #[\NoDiscard]
     public function format(): string
     {
         if ($this->value === '') {
@@ -64,6 +67,7 @@ final class Query
         return '?'.$this->value;
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value;