+++ tests/Authority/UserInformationTest.php
@@ -7,9 +7,8 @@ use Innmind\Url\{
     Authority\UserInformation,
     Authority\UserInformation\User,
     Authority\UserInformation\Password,
-    Exception\PasswordCannotBeSpecifiedWithoutAUser,
 };
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class UserInformationTest extends TestCase
 {
@@ -28,9 +27,9 @@ class UserInformationTest extends TestCase
 
     public function testThrowWhenNullUserButPasswordPresent()
     {
-        $this->expectException(PasswordCannotBeSpecifiedWithoutAUser::class);
+        $this->expectException(\DomainException::class);
 
-        UserInformation::of(User::none(), Password::of('foo'));
+        $_ = UserInformation::of(User::none(), Password::of('foo'));
     }
 
     public function testString()