+++ tests/SchemeTest.php
@@ -3,11 +3,8 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Tests;
 
-use Innmind\Url\{
-    Scheme,
-    Exception\DomainException,
-};
-use PHPUnit\Framework\TestCase;
+use Innmind\Url\Scheme;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class SchemeTest extends TestCase
 {
@@ -21,10 +18,10 @@ class SchemeTest extends TestCase
 
     public function testThrowWhenInvalidData()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('http://');
 
-        Scheme::of('http://');
+        $_ = Scheme::of('http://');
     }
 
     public function testNull()